+++ lang/en/filter_filtercodes.php
@@ -49,6 +49,8 @@ $string['courseteacherlinktype'] = 'Teacher link type';
 $string['courseteacherlinktype_desc'] = 'Choose the type of link for the teacher\s link in the {courseteachers} tags.';
 $string['ifprofilefiedonlyvisible'] = '{ifprofile_field_} only visible.';
 $string['ifprofilefiedonlyvisible_desc'] = 'Restrict the {ifprofile_field_...} tag to only access visible profile fields. Hidden fields will behave as if the field was empty. If unchecked, this tag will be able to check hidden user fields.';
+$string['categorycardshowpic'] = 'Show {categorycards} background';
+$string['categorycardshowpic_desc'] = 'If enabled, will display a background colour/pattern for {categorycards} tag similar to the course images when no image was specified.';
 
 $string['sizeb'] = 'B';
 $string['sizekb'] = 'KB';