+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [2.2.2] 2021-05-26 (DEV)
+## [2.2.2] 2021-06-01 (DEV)
 ### Added
 - New {ifnothome}{/ifnothome} tags.
-
+- New setting to enable/disable {categorycards} background.
+- New fc-categorycards class to {categorycards} to make it easier to customize.
 ## [2.2.1] 2021-05-25
 ### Updated
 - Corrections in some of the language strings.